address comments from Dirk
